fixed calculator and invoice/ added responsibility

+++ src/script/script.js
@@ -1,6 +1,7 @@
 // import
 import { getProducts } from "../script/ApiServices/apiService.js";
 
+
 // fetching by using axios
 async function fetchData() {
   try {
@@ -20,7 +21,7 @@ function displayData(data) {
   itemDiv.innerHTML = data
     .map(
       (item) => `
-        <div class="p-3 itemCard pb-4 d-flex flex-column gap-2">
+        <div class="p-3 itemCard pb-4 d-flex col-5 col-md-3 flex-column gap-2">
           <img src="${item.image}">
           <p class="title">${item.name}</p>
           <div class="id">
@@ -34,6 +35,13 @@ function displayData(data) {
   const itemCards = itemDiv.getElementsByClassName("itemCard");
   Array.from(itemCards).forEach((itemCard, index) => {
     itemCard.addEventListener("click", () => {
+      Swal.fire({
+        position: 'top-end',
+        icon: 'success',
+        title: 'Item is added to the cart',
+        showConfirmButton: false,
+        timer: 1500
+      })
       const selectedItem = data[index];
 
       let selectedItems = JSON.parse(localStorage.getItem("cartItems")) || [];
@@ -44,10 +52,10 @@ function displayData(data) {
         alert("Item is already added to the cart.");
         return;
       }
-      selectedItem.quantity = 1; // Set initial quantity
+      selectedItem.quantity = 1; 
       selectedItems.push(selectedItem);
       localStorage.setItem("cartItems", JSON.stringify(selectedItems));
-      displayCartItems(data); // Pass 'data' as a parameter
+      displayCartItems(data); 
     });
   });
 
@@ -127,7 +135,7 @@ function displayCartItems(data) {
             <input id="discountInput" type="number" min="0" step="0.01" value="0">
           </div>
           <div class="d-flex flex-column col-4">
-            <label>Shipping %</label>
+            <label>Shipping:</label>
             <input id="shippingInput" type="number" min="0" step="0.01" value="0">
           </div>
         </div>
@@ -135,7 +143,7 @@ function displayCartItems(data) {
           <h1 class="col-8">Total price: <span id="totalPrice">0</span>$</h1>
           <div class="d-flex justify-content-around col-10 mx-auto">
             <button id="resetBtn"><i class="fa-solid fa-power-off"></i> Reset</button>
-            <button><i class="fa-solid fa-file-invoice-dollar fa-sm"></i> Invoice</button>
+            <button id="invoice"><i class="fa-solid fa-file-invoice-dollar fa-sm"></i> Invoice</button>
           </div>
         </footer>
       </div>
@@ -145,10 +153,28 @@ function displayCartItems(data) {
     const discountInput = document.getElementById("discountInput");
     const shippingInput = document.getElementById("shippingInput");
     const totalPriceSpan = document.getElementById("totalPrice");
+    const invoiceBTN = document.getElementById('invoice')
 
     resetBtn.addEventListener("click", () => {
-      localStorage.removeItem("cartItems");
-      displayCartItems(data);
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
+          localStorage.removeItem("cartItems");
+        displayCartItems(data);
+        }
+      })
     });
 
     const ChoosedItems = document.getElementById("ChoosedItems");
@@ -157,17 +183,17 @@ function displayCartItems(data) {
     cartItems.forEach((item) => {
       const subtotal = item.price * item.quantity;
       subtotalsSum += subtotal;
-      calculateTotalPrice()
+      calculateTotalPrice();
 
       ChoosedItems.innerHTML += `
         <header class="d-flex gap-3 justify-content-start">
-          <div class="col-2">
+          <div class="col-3">
             <p>${item.name} #${item.id}</p>
           </div>
-          <div class="col-1">
+          <div class="col-2">
             <p>${item.price}$</p>
           </div>
-          <div class="col-3 d-flex gap-2">
+          <div class="col-2 d-flex gap-2">
             <button class="decrement-btn">-</button>
             <div class="qty">${item.quantity}</div>
             <button class="increment-btn">+</button>
@@ -190,9 +216,26 @@ function displayCartItems(data) {
 
     deleteBtns.forEach((deleteButton, index) => {
       deleteButton.addEventListener("click", () => {
-        cartItems.splice(index, 1);
-        localStorage.setItem("cartItems", JSON.stringify(cartItems));
-        displayCartItems(data);
+        Swal.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+            cartItems.splice(index, 1);
+            localStorage.setItem("cartItems", JSON.stringify(cartItems));
+            displayCartItems(data);
+          }
+        })
       });
     });
 
@@ -226,16 +269,45 @@ function displayCartItems(data) {
 
     function calculateTotalPrice() {
       const tax = parseFloat(taxInput.value);
+      localStorage.setItem('tax', tax)
       const discount = parseFloat(discountInput.value);
+      localStorage.setItem('discount', discount)
       const shipping = parseFloat(shippingInput.value);
-      const subtotalsSum = cartItems.reduce((sum, item) => sum + item.price * item.quantity, 0);
-      const total = subtotalsSum * (1 + tax / 100) * (1 - discount / 100) + shipping;
-      totalPriceSpan.innerText = total.toFixed(1);
-    
-      if (tax.length === 0 || discount.length === 0 || shipping.length === 0) {
-        totalPriceSpan.innerText = subtotalsSum.toFixed(1);
-      } 
-
+      localStorage.setItem('shipping', shipping)
+      const subtotalsSum = cartItems.reduce(
+        (sum, item) => sum + item.price * item.quantity,
+        0
+      );
+      const total =
+        subtotalsSum * (1 + tax / 100) * (1 - discount / 100) + shipping;
+        totalPriceSpan.innerText = total.toFixed(1);
+        localStorage.setItem('sum',  total.toFixed(1))
+      if (isNaN(tax) || isNaN(discount) || isNaN(shipping)) {
+        totalPriceSpan.innerText = subtotalsSum.toFixed(1); 
+      }
+      invoiceBTN.addEventListener('click', () => {
+        const sum = localStorage.getItem('sum');
+        const tax = localStorage.getItem('tax');
+        const discount = localStorage.getItem('discount');
+        const shipping = localStorage.getItem('shipping');
+      
+        Swal.fire({
+          title: 'Invoice',
+          html: `
+            <p>Total: ${sum}$</p>
+            <p>Tax: ${tax}%</p>
+            <p>Discount: ${discount}%</p>
+            <p>Shipping: ${shipping}$</p>
+          `,
+          icon: 'info',
+          confirmButtonText: 'PURCHASE'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire('Purchase has been complited!', '', 'success')
+          }
+        });
+      });
+      
     }
   } else {
     const cartItemsDiv = document.getElementById("cartItems");
